gralloc_gbm_mesa: Fix -Wcast-qual and -Wcast-align errors in gralloc_handle()
The module is built with -Wcast-qual -Wcast-align, and the pointer cast
in gralloc_handle() violates both:

  * buffer_handle_t is a 'const native_handle_t *', so casting it to a
    non-const 'struct gralloc_handle_t *' drops the const qualifier.
  * native_handle_t only requires 4-byte alignment, while
    gralloc_handle_t requires 8-byte alignment because of its
    aligned(8) modifier field and union.

Round-trip the pointer through uintptr_t instead. Neither warning
applies to pointer-to-integer-to-pointer conversions, and uintptr_t is
exactly pointer sized so -Wint-to-pointer-cast does not fire either.
This keeps the header valid C, unlike a const_cast/static_cast chain,
so it stays a drop-in copy of external/libdrm/android/gralloc_handle.h.

The cast is safe in practice: the storage always comes from
native_handle_create(), which uses malloc() and therefore returns
memory aligned for any type.

+++ src/include/drm/gralloc_handle.h
@@ -78,7 +78,16 @@ struct gralloc_handle_t {
 
 static inline struct gralloc_handle_t *gralloc_handle(buffer_handle_t handle)
 {
-	return (struct gralloc_handle_t *)handle;
+	/*
+	 * buffer_handle_t is a 'const native_handle_t *', which is only
+	 * 4-byte aligned, while gralloc_handle_t requires 8-byte alignment.
+	 * The underlying buffer is always allocated by native_handle_create(),
+	 * i.e. by malloc(), so it is suitably aligned for gralloc_handle_t.
+	 * Round-tripping through uintptr_t drops the const qualifier and the
+	 * alignment assumptions in a single step, which keeps both
+	 * -Wcast-qual and -Wcast-align quiet in C as well as in C++.
+	 */
+	return (struct gralloc_handle_t *)(uintptr_t)handle;
 }
 
 /**
